requirements.txt
requests
duckdb
python-dotenv-vault
